+++ src/parser/api/source.ts
@@ -1,0 +1,31 @@
+import { BlockParser } from '../block';
+import { bind } from '../../combinator';
+import { parse } from './parse';
+
+const sources: WeakMap<Element, string> = new WeakMap();
+
+export function breaklines(source: string): string {
+  return [...parse(source).children]
+    .map<string>(el => {
+      if (el instanceof HTMLParagraphElement === false) return sources.get(el)!;
+      const breaks = el.querySelectorAll('br, .linebreak');
+      return sources.get(el)!
+        .split('\n')
+        .map((line, i) =>
+          breaks[i] && breaks[i].matches('.linebreak')
+            ? `${line}\\`
+            : line)
+        .join('\n');
+    })
+    .join('\n');
+}
+
+export function memorize(parser: BlockParser): BlockParser {
+  return ((source: string) =>
+    bind(parser, (rs, rest) => {
+      if (rs.length === 0) return [rs, rest];
+      assert(rs.length === 1);
+      void sources.set(rs[0], source.slice(0, source.length - rest.length));
+      return [rs, rest];
+    })(source)) as BlockParser;
+}